requirements.txt
# requirements.txt
ccxt>=2.0.0
pandas>=1.3.0
numpy>=1.20.0
pyyaml>=6.0